Fix query examples on Readme.md and MacOSx install tip for MAMP
queries needed double \ to escape. The error may have been due to
markdown's escaping \\ into \.
also added NOT NULL condition to reduce possible confusion that may
cause users to think there is an error in the library.
without CONVERT, PREG_REPLACE returns a binary result, Mysql Workbench
displays as if the results are empty, which is highly confusing.